branches/branches 64 bits, work in progress
git-svn-id: https://svn.lcsr.jhu.edu/cisst@740 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ branches/branches/2009-08-19-64bits/libs/code/cisstCommon/cmnRandomSequence.cpp
@@ -29,9 +29,9 @@ const cmnRandomSequence::ElementaryRandomNumber cmnRandomSequence::UpperRandomBo
 
 cmnRandomSequence cmnRandomSequence::RandomInstance;
 
-void cmnRandomSequence::ExtractRandomPermutation(const unsigned int length, unsigned int * array)
+void cmnRandomSequence::ExtractRandomPermutation(const size_t length, unsigned int * array)
 {
-    unsigned int i;
+    size_t i;
     for (i = 0; i < length; ++i)
         array[i] = ExtractRandomInt(0, length);
     
@@ -46,9 +46,9 @@ void cmnRandomSequence::ExtractRandomPermutation(const unsigned int length, unsi
 }
 
 
-void cmnRandomSequence::ExtractRandomPermutation(const unsigned long long int length, unsigned long long int * array)
+void cmnRandomSequence::ExtractRandomPermutation(const size_t length, unsigned long long int * array)
 {
-    unsigned long long i;
+    size_t i;
     for (i = 0; i < length; ++i)
         array[i] = ExtractRandomUnsignedLongLong(0, length);
